refactor(#1235): separate atomic constraint reading into 3 sections
