+++ profile/src/main/java/com/scottlogic/deg/profile/reader/atomic/AtomicConstraintFactory.java
@@ -7,15 +7,10 @@ import com.scottlogic.deg.common.profile.constraints.Constraint;
 import com.scottlogic.deg.common.profile.constraints.atomic.*;
 import com.scottlogic.deg.common.util.NumberUtils;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.DistributedSet;
-import com.scottlogic.deg.generator.fieldspecs.whitelist.FrequencyDistributedSet;
 import com.scottlogic.deg.profile.dto.AtomicConstraintType;
-import com.scottlogic.deg.profile.reader.ConstraintReaderHelpers;
 import com.scottlogic.deg.profile.reader.RemoveFromTree;
 
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
-import java.util.Collection;
-import java.util.Map;
 import java.util.regex.Pattern;
 
 public class AtomicConstraintFactory {
