+++ profile/src/main/java/com/scottlogic/deg/profile/reader/atomic/AtomicConstraintValueReader.java
@@ -1,12 +1,12 @@
 package com.scottlogic.deg.profile.reader.atomic;
 
 import com.google.inject.Inject;
-import com.scottlogic.deg.common.util.NumberUtils;
+import com.scottlogic.deg.common.ValidationException;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.FrequencyDistributedSet;
+import com.scottlogic.deg.profile.dto.AtomicConstraintType;
 import com.scottlogic.deg.profile.dto.ConstraintDTO;
-import com.scottlogic.deg.profile.reader.ConstraintReaderHelpers;
+import com.scottlogic.deg.profile.reader.InvalidProfileException;
 
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.Collection;
 import java.util.List;
@@ -22,39 +22,44 @@ public class AtomicConstraintValueReader {
         this.fromFileReader = fromFileReader;
     }
 
+
     public Object getValue(ConstraintDTO dto){
+        try {
+            return tryGetValue(dto);
+        } catch (IllegalArgumentException | ValidationException e){
+            throw new InvalidProfileException(String.format("Field [%s]: %s", dto.field, e.getMessage()));
+        }
+    }
+
+    public Object tryGetValue(ConstraintDTO dto){
         if (dto.values != null){
-            return getSet(dto.values, dto);
+            return getSet(dto.values);
         }
 
         if (dto.file != null){
             return fromFileReader.setFromFile(dto.file);
         }
 
-        return getValue(dto.value, dto);
+        return getValue(dto.value);
     }
 
-    private FrequencyDistributedSet getSet(Collection<Object> values, ConstraintDTO dto) {
-        if (!(values instanceof Collection)){
-            return null;//TODO lul
-        }
-
+    private FrequencyDistributedSet getSet(Collection<Object> values) {
         List collect = values.stream()
-            .map(val -> getValue(val, dto))
+            .map(val -> getValue(val))
             .collect(Collectors.toList());
         return FrequencyDistributedSet.uniform(collect);
     }
 
-    private Object getValue(Object value, ConstraintDTO dto) {
+    private Object getValue(Object value) {
         if (value instanceof Map){
-            return getDate((Map) value, dto);
+            return getDate((Map) value);
         }
 
         return value;
     }
 
 
-    private OffsetDateTime getDate(Map value, ConstraintDTO dto) {
-        return ConstraintReaderHelpers.parseDate((String) (value).get("date"), dto);
+    private OffsetDateTime getDate(Map value) {
+        return ConstraintReaderHelpers.parseDate((String) (value).get("date"));
     }
 }
